requirements.txt
streamlit>=1.42.0
pandas>=2.2.0
pyyaml>=6.0.1
watchdog>=6.0.0
pytest>=8.3.4
pytest-cov>=4.1.0
psutil>=5.9.8
